+++ docs/notes/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Next Version
+------------
+
+**New Features**
+
+* Introduced :class:`evalmate.tasks.Outcome` to have a common input structure for reference and hypothesis.
+
+* With :class:`evalmate.tasks.LabelSet` more statistics on reference and hypothesis can be computed.
+  Label-Sets are created via :class:`evalmate.tasks.Outcome` class.
+
 v0.1.0
 ------